week4_web_implementation/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
PyJWT
